iroh-ble-transport: retire the routing_v2 name, fold helpers into one module
With v1 gone (step 6), the `_v2` suffix is vestigial. Rename
`routing_v2.rs` → `routing.rs` and fold the handful of stateless
address-layer helpers (previously a separate 90-line `routing.rs`:
`prefix_from_endpoint`, `token_custom_addr`, `parse_token_addr`,
`TOKEN_LEN`) into the same module — they're all consumers of
`StableConnId` and were already cross-importing.

* One module (`crate::transport::routing`) owns the routing table,
  the `CustomAddr` wire format, and the `EndpointId → KeyPrefix`
  projection. Doc comment at the top of the file documents both
  sections.
* All 112 `transport::routing_v2::*` call sites rewrite to
  `transport::routing::*` — mechanical sed-style rename, no
  behavior change. The existing `transport::routing::prefix_from_endpoint`
  etc. references stay exactly the same.
* `pub mod routing_v2` removed from `transport/mod.rs`.

318 tests pass, clippy clean.
