+++ crates/iroh-ble-transport/src/transport/l2cap.rs
@@ -72,7 +72,7 @@ pub(super) fn spawn_l2cap_io_tasks<R, W>(
     reader: R,
     writer: W,
     device_id: blew::DeviceId,
-    stable_conn_id: crate::transport::routing_v2::StableConnId,
+    stable_conn_id: crate::transport::routing::StableConnId,
     incoming_tx: mpsc::Sender<IncomingPacket>,
     last_rx_at: crate::transport::peer::LivenessClock,
     tearing_down: Arc<AtomicBool>,
@@ -261,7 +261,7 @@ mod tests {
             a_rd,
             a_wr,
             device_id.clone(),
-            crate::transport::routing_v2::StableConnId::for_test(42),
+            crate::transport::routing::StableConnId::for_test(42),
             incoming_tx,
             crate::transport::peer::LivenessClock::new(),
             Arc::new(AtomicBool::new(false)),
@@ -300,7 +300,7 @@ mod tests {
             a_rd,
             a_wr,
             device_id.clone(),
-            crate::transport::routing_v2::StableConnId::for_test(42),
+            crate::transport::routing::StableConnId::for_test(42),
             incoming_tx,
             crate::transport::peer::LivenessClock::new(),
             Arc::new(AtomicBool::new(false)),
@@ -334,7 +334,7 @@ mod tests {
             a_rd,
             a_wr,
             blew::DeviceId::from("exit-test"),
-            crate::transport::routing_v2::StableConnId::for_test(42),
+            crate::transport::routing::StableConnId::for_test(42),
             incoming_tx,
             crate::transport::peer::LivenessClock::new(),
             Arc::new(AtomicBool::new(false)),
@@ -361,7 +361,7 @@ mod tests {
             a_rd,
             a_wr,
             device_id.clone(),
-            crate::transport::routing_v2::StableConnId::for_test(42),
+            crate::transport::routing::StableConnId::for_test(42),
             incoming_tx,
             crate::transport::peer::LivenessClock::new(),
             Arc::new(AtomicBool::new(false)),
@@ -399,7 +399,7 @@ mod tests {
             a_rd,
             a_wr,
             blew::DeviceId::from("abort-test"),
-            crate::transport::routing_v2::StableConnId::for_test(42),
+            crate::transport::routing::StableConnId::for_test(42),
             incoming_tx,
             crate::transport::peer::LivenessClock::new(),
             Arc::new(AtomicBool::new(false)),
