+++ crates/iroh-ble-transport/src/transport/mod.rs
@@ -13,7 +13,6 @@ pub mod pipe_watchdog;
 pub mod registry;
 pub mod reliable;
 pub mod routing;
-pub mod routing_v2;
 pub mod store;
 #[allow(clippy::module_inception)]
 pub mod transport;
