+++ crates/iroh-ble-transport/tests/integration_dedup.rs
@@ -82,7 +82,7 @@ fn spawn_node_with_endpoint(
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
 
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         Arc::clone(&iface),
         inbox_tx.clone(),
@@ -91,14 +91,14 @@ fn spawn_node_with_endpoint(
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let registry = Registry::new_for_test_with_policy_and_endpoint(policy, endpoint);
     let snap = snapshots.clone();
     let wakers = Arc::new(Mutex::new(Vec::<Waker>::new()));
     tokio::spawn(async move {
         registry
-            .run(inbox_rx, driver, snap, wakers, routing_v2_local)
+            .run(inbox_rx, driver, snap, wakers, routing_local)
             .await;
     });
