+++ crates/iroh-ble-transport/tests/mock_integration.rs
@@ -64,7 +64,7 @@ async fn mid_session_disconnect_drains_and_closes_channel() {
     let (incoming_tx, _incoming_rx) = mpsc::channel::<IncomingPacket>(16);
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         iface.clone(),
         tx.clone(),
@@ -73,7 +73,7 @@ async fn mid_session_disconnect_drains_and_closes_channel() {
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let reg = Registry::new_for_test_with_policy(L2capPolicy::Disabled);
     let snap_for_actor = snapshots.clone();
@@ -83,7 +83,7 @@ async fn mid_session_disconnect_drains_and_closes_channel() {
             driver,
             snap_for_actor,
             test_wakers(),
-            Arc::clone(&routing_v2_local),
+            Arc::clone(&routing_local),
         )
         .await;
     });
@@ -180,7 +180,7 @@ async fn adapter_toggle_reconnects_all_peers_with_single_purge() {
     let (incoming_tx, _incoming_rx) = mpsc::channel::<IncomingPacket>(16);
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         iface.clone(),
         tx.clone(),
@@ -189,7 +189,7 @@ async fn adapter_toggle_reconnects_all_peers_with_single_purge() {
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let reg = Registry::new_for_test_with_policy(L2capPolicy::Disabled);
     let snap_for_actor = snapshots.clone();
@@ -199,7 +199,7 @@ async fn adapter_toggle_reconnects_all_peers_with_single_purge() {
             driver,
             snap_for_actor,
             test_wakers(),
-            Arc::clone(&routing_v2_local),
+            Arc::clone(&routing_local),
         )
         .await;
     });
@@ -258,7 +258,7 @@ async fn adapter_on_rebuilds_server_and_restarts_advertising_and_l2cap() {
     let (incoming_tx, _incoming_rx) = mpsc::channel::<IncomingPacket>(16);
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         iface.clone(),
         tx.clone(),
@@ -267,7 +267,7 @@ async fn adapter_on_rebuilds_server_and_restarts_advertising_and_l2cap() {
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let reg = Registry::new_for_test_with_policy(L2capPolicy::PreferL2cap);
     let snap_for_actor = snapshots.clone();
@@ -277,7 +277,7 @@ async fn adapter_on_rebuilds_server_and_restarts_advertising_and_l2cap() {
             driver,
             snap_for_actor,
             test_wakers(),
-            Arc::clone(&routing_v2_local),
+            Arc::clone(&routing_local),
         )
         .await;
     });
@@ -332,7 +332,7 @@ async fn version_mismatch_deads_peer_and_closes_channel() {
     let (incoming_tx, _incoming_rx) = mpsc::channel::<IncomingPacket>(8);
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         iface.clone(),
         tx.clone(),
@@ -341,7 +341,7 @@ async fn version_mismatch_deads_peer_and_closes_channel() {
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let reg = Registry::new_for_test_with_policy(L2capPolicy::Disabled);
     let snap_for_actor = snapshots.clone();
@@ -351,7 +351,7 @@ async fn version_mismatch_deads_peer_and_closes_channel() {
             driver,
             snap_for_actor,
             test_wakers(),
-            Arc::clone(&routing_v2_local),
+            Arc::clone(&routing_local),
         )
         .await;
     });
@@ -429,7 +429,7 @@ async fn version_match_lets_data_pipe_start() {
     let (incoming_tx, _incoming_rx) = mpsc::channel::<IncomingPacket>(8);
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         iface.clone(),
         tx.clone(),
@@ -438,7 +438,7 @@ async fn version_match_lets_data_pipe_start() {
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let reg = Registry::new_for_test_with_policy(L2capPolicy::Disabled);
     let snap_for_actor = snapshots.clone();
@@ -448,7 +448,7 @@ async fn version_match_lets_data_pipe_start() {
             driver,
             snap_for_actor,
             test_wakers(),
-            Arc::clone(&routing_v2_local),
+            Arc::clone(&routing_local),
         )
         .await;
     });
@@ -512,7 +512,7 @@ async fn mock_fabric_handles_bidirectional_traffic() {
         c_truncations,
         c_empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new()),
+        Arc::new(iroh_ble_transport::transport::routing::Routing::new()),
     );
     let peripheral_driver = Driver::new(
         fabric.peripheral.clone(),
@@ -522,7 +522,7 @@ async fn mock_fabric_handles_bidirectional_traffic() {
         p_truncations,
         p_empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new()),
+        Arc::new(iroh_ble_transport::transport::routing::Routing::new()),
     );
 
     let central_reg = Registry::new_for_test_with_policy(L2capPolicy::Disabled);
@@ -536,7 +536,7 @@ async fn mock_fabric_handles_bidirectional_traffic() {
                 central_driver,
                 cs,
                 test_wakers(),
-                Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new()),
+                Arc::new(iroh_ble_transport::transport::routing::Routing::new()),
             )
             .await;
     });
@@ -547,7 +547,7 @@ async fn mock_fabric_handles_bidirectional_traffic() {
                 peripheral_driver,
                 ps,
                 test_wakers(),
-                Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new()),
+                Arc::new(iroh_ble_transport::transport::routing::Routing::new()),
             )
             .await;
     });
@@ -669,7 +669,7 @@ async fn forget_then_gc_then_rediscover_creates_fresh_peer() {
     let (incoming_tx, _incoming_rx) = mpsc::channel::<IncomingPacket>(16);
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         iface.clone(),
         tx.clone(),
@@ -678,7 +678,7 @@ async fn forget_then_gc_then_rediscover_creates_fresh_peer() {
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let reg = Registry::new_for_test_with_policy(L2capPolicy::Disabled);
     let snap_for_actor = snapshots.clone();
@@ -688,7 +688,7 @@ async fn forget_then_gc_then_rediscover_creates_fresh_peer() {
             driver,
             snap_for_actor,
             test_wakers(),
-            Arc::clone(&routing_v2_local),
+            Arc::clone(&routing_local),
         )
         .await;
     });
@@ -793,7 +793,7 @@ async fn connect_failure_retries_on_next_tick() {
     let (incoming_tx, _incoming_rx) = mpsc::channel::<IncomingPacket>(16);
     let snapshots = Arc::new(ArcSwap::from(Arc::new(SnapshotMaps::default())));
     let (retransmits, truncations, empty_frames) = zero_counters();
-    let routing_v2_local = Arc::new(iroh_ble_transport::transport::routing_v2::Routing::new());
+    let routing_local = Arc::new(iroh_ble_transport::transport::routing::Routing::new());
     let driver = Driver::new(
         iface.clone(),
         tx.clone(),
@@ -802,7 +802,7 @@ async fn connect_failure_retries_on_next_tick() {
         truncations,
         empty_frames,
         Arc::new(InMemoryPeerStore::new()),
-        Arc::clone(&routing_v2_local),
+        Arc::clone(&routing_local),
     );
     let reg = Registry::new_for_test_with_policy(L2capPolicy::Disabled);
     let snap_for_actor = snapshots.clone();
@@ -812,7 +812,7 @@ async fn connect_failure_retries_on_next_tick() {
             driver,
             snap_for_actor,
             test_wakers(),
-            Arc::clone(&routing_v2_local),
+            Arc::clone(&routing_local),
         )
         .await;
     });
